+++ src/components/common/header/UserModal.tsx
@@ -2,13 +2,13 @@ import logoutIcon from '@/assets/icons/logout.svg'
 import modalProfileIcon from '@/assets/icons/modalProfileIcon.svg'
 function UserModal() {
   return (
-    <div className="hidden md:absolute md:top-[45.05px] md:right-[10px] md:flex md:h-[99px] md:w-[192px] md:flex-col md:gap-[8px] md:rounded-lg md:border md:border-solid md:border-[#E5E7EB] md:bg-[#ffffff] md:drop-shadow-[0_10px_15px_rgba(0,0,0,0.2)]">
-      <div className="mt-[13px] flex cursor-pointer items-center gap-[12px] border-t-[2px] border-solid border-[#f3f4f6] px-[16px] py-[10px] pt-[5px] text-[16px] text-[#374151]">
+    <div className="md:bg-basic-white hidden md:absolute md:top-[45.05px] md:right-2.5 md:flex md:h-[99px] md:w-48 md:flex-col md:gap-2 md:rounded-lg md:border md:border-solid md:border-[#E5E7EB] md:drop-shadow-[0_10px_15px_rgba(0,0,0,0.2)]">
+      <div className="mt-[13px] flex cursor-pointer items-center gap-3 border-t-2 border-solid border-gray-100 px-4 py-2.5 pt-[5px] text-[16px] text-gray-700">
         <img src={logoutIcon} alt="logoutIcon" className="h-[18px] w-[18px]" />
         <span>로그아웃</span>
         {/* 클릭하면 로그아웃 시키기 */}
       </div>
-      <div className="py-[10px]text-[16px] flex cursor-pointer items-center gap-[12px] px-[16px] text-[#374151]">
+      <div className="flex cursor-pointer items-center gap-3 px-4 py-2.5 text-[16px] text-gray-700">
         <img
           src={modalProfileIcon}
           alt="modalProfileIcon"
